Split stack state into multiple files
